% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add.GeneList}
\alias{add.GeneList}
\title{Add a gene list to an scTypeEval object.}
\usage{
add.GeneList(scTypeEval, gene.list = NULL)
}
\arguments{
\item{scTypeEval}{An \code{scTypeEval} object to which the gene list will be added.}

\item{gene.list}{A named list of gene sets to append. If the list is unnamed, names will be assigned automatically.}
}
\value{
An updated \code{scTypeEval} object with the new gene list added.
}
\description{
This function appends a new gene list to an existing \code{scTypeEval} object.
It ensures that the input list is valid and assigns names if they are missing.
}
\details{
The function verifies that \code{gene.list} is provided and is a valid list.
If any elements in \code{gene.list} lack names, they are automatically renamed.
}
\examples{
\dontrun{
sceval <- add.GeneList(sceval, gene.list = list("cytokines" = c("IL10", "IL6", "IL4",...)))
}

}
