% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.PCA}
\alias{plot.PCA}
\title{Plot PCA Results from scTypeEval Object}
\usage{
\method{plot}{PCA}(
  scTypeEval,
  gene.list = NULL,
  data.type = NULL,
  dims = c(1, 2),
  label = TRUE,
  show.legend = F
)
}
\arguments{
\item{scTypeEval}{An \code{scTypeEval} object containing PCA results in the \code{reductions} slot.}

\item{gene.list}{Character vector. A subset of genes to filter PCA plots. If \code{NULL}, all available PCA results are plotted.}

\item{data.type}{Character vector. Specifies which data types to include in the plots (e.g., \code{"sc"}, \code{"pseudobulk"}).
If \code{NULL}, all data types are included.}

\item{dims}{Integer vector of length 2. The principal component (PC) dimensions to plot (default: \code{c(1,2)}).}

\item{label}{Logical. Whether to add labels to the PCA plot (default: \code{TRUE}).}

\item{show.legend}{Logical. Whether to display a legend (default: \code{FALSE}).}
}
\value{
A named list of PCA plots (\link{ggplot2} objects) corresponding to different PCA analyses stored in
\code{reductions} slot of a scTypeEval object by \link{add.PCA}.
}
\description{
This function visualizes Principal Component Analysis (PCA) results stored in the \code{reductions} slot of a scTypeEval object.
}
\examples{
\dontrun{
# plot all PCA dimensional reductions
pca_plots <- plot.PCA(sceval) # scTypeEval object with previously run `add.PCA()`

# plot only some PCAs
pca_plots_filtered <- plot.PCA(sceval, # scTypeEval object with previously run `add.PCA()`
                              gene.list = "HVG", # plot only PCA for highly variable genes
                              data.type = "pseudobulk" # plot only PCA for pseudobulk data
                              ) 
}

}
\seealso{
\link{add.PCA}
}
