% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA Results from scTypeEval Object}
\usage{
plot_pca(
  scTypeEval,
  reduction_slot = "all",
  label = TRUE,
  dims = c(1, 2),
  show_legend = FALSE
)
}
\arguments{
\item{scTypeEval}{An \code{scTypeEval} object containing PCA results in the \code{reductions} slot.}

\item{reduction_slot}{Character. Name(s) of the reduction(s) to plot. If \code{"all"}
(default), all available PCA reductions in the object are plotted.}

\item{label}{Logical. Whether to add cluster labels to the PCA plot (default: \code{TRUE}).}

\item{dims}{Integer vector of length 2. The principal component (PC) dimensions to plot
(default: \code{c(1,2)}).}

\item{show_legend}{Logical. Whether to display a legend (default: \code{FALSE}).}
}
\value{
A named list of PCA plots (\link[ggplot2]{ggplot} objects) corresponding to
the PCA analyses stored in the \code{reductions} slot of an \code{scTypeEval} object
(generated by \link{add_dim_reduction}). If only one reduction is selected, a single ggplot object
is returned.
}
\description{
This function visualizes Principal Component Analysis (PCA) results stored in the
\code{reductions} slot of an \code{scTypeEval} object.
}
\examples{
# Create and process test data
library(Matrix)
counts <- Matrix(rpois(6000, 5), nrow = 100, ncol = 60, sparse = TRUE)
rownames(counts) <- paste0("Gene", 1:100)
colnames(counts) <- paste0("Cell", 1:60)

metadata <- data.frame(
  celltype = rep(c("TypeA", "TypeB"), each = 30),
  sample = rep(paste0("Sample", 1:6), times = 10),
  row.names = colnames(counts)
)

sceval <- create_scTypeEval(matrix = counts, metadata = metadata)
sceval <- run_processing_data(sceval, ident = "celltype",
                             sample = "sample", min_samples = 3,
                             min_cells = 3, verbose = FALSE)
sceval <- run_hvg(sceval,
                  var_method = "basic",
                  ngenes = 20,
                  verbose = FALSE)

# Run PCA on HVG genes
sceval <- run_pca(sceval,
                 gene_list = "HVG",
                 ndim = 4,
                 verbose = FALSE)

# plot PCA
plot_pca(sceval)

}
\seealso{
\link{add_dim_reduction}
}
