% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Run.scTypeEval}
\alias{Run.scTypeEval}
\title{Wrapper for Consistency Metrics and BestHit Evaluation for multiple data.type inputs.}
\usage{
Run.scTypeEval(
  scTypeEval,
  ident = NULL,
  ident_GroundTruth = NULL,
  sample = NULL,
  normalization.method = c("Log1p", "CLR", "pearson"),
  gene.list = NULL,
  pca = FALSE,
  ndim = 30,
  distance.method = "euclidean",
  IntVal.metric = c("silhouette", "NeighborhoodPurity", "ward.PropMatch",
    "Leiden.PropMatch", "modularity"),
  BH.method = c("Mutual.Score", "Mutual.Match"),
  data.type = c("sc", "pseudobulk", "pseudobulk_1vsall"),
  min.samples = 5,
  min.cells = 10,
  k.sc = 20,
  k.psblk = 5,
  black.list = NULL,
  ncores = 1,
  bparam = NULL,
  progressbar = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{scTypeEval}{A scTypeEval object containing single-cell expression data, metadata, and gene lists.}

\item{ident}{A character string specifying the metadata column containing the cell type annotations.
Default is \code{NULL}, which uses the active identity in \code{scTypeEval}.}

\item{ident_GroundTruth}{(Optional) A character string specifying the metadata column containing the
ground truth annotations. Default is \code{NULL}, meaning \code{ident} is used. See \link[Run.BestHit]{Run.BestHit} for details.}

\item{sample}{A character string specifying the metadata column that identifies sample IDs.
Required for consistency evaluation when using pseudobulk data. For \code{"sc"} data.type \code{sample == NULL} is considered, so inter-sample variability is not accounted for.}

\item{normalization.method}{Character. Method for normalizing the expression data. See \link[add.HVG]{add.HVG} for more details.
Options: \code{"Log1p"}, \code{"CLR"}, \code{"pearson"}. Default: \code{"Log1p"}.}

\item{gene.list}{List. A list of gene sets to compute consistency metrics on.
By default, each of the gene lists stored in \code{scTypeEval} are used recursively to compute consistency metrics.}

\item{pca}{Logical. Whether to perform PCA before consistency evaluation. Default: \code{FALSE}. Only applies to \link[Run.Consistency]{Run.Consistency}.}

\item{ndim}{Integer. Number of principal components to retain if \code{pca = TRUE}. Default: \code{30}. Only applies to \link[Run.Consistency]{Run.Consistency}.}

\item{distance.method}{Character. Method to compute distances between samples when running \link[Run.Consistency]{Run.Consistency}. Default is \code{"euclidean"}.
See \link[Run.Consistency]{Run.Consistency} for details.}

\item{IntVal.metric}{A character vector specifying internal validation metrics to compute. See \link[Run.Consistency]{Run.Consistency} for details.}

\item{BH.method}{A character vector specifying Mutual Best Hit methods.
Options: \code{"Mutual.Score"} and \code{"Mutual.Match"}, by default both are run if posssible.
See \link[Run.BestHit]{Run.BestHit} for details.}

\item{data.type}{A character vector specifying the type(s) of data input. If multiple are provided, all of them will be run recursively if supported.
By default all are run. Options:
\itemize{
\item \code{"sc"}
\item \code{"pseudobulk"}
\item \code{"pseudobulk_1vsall"}
}}

\item{min.samples}{Integer. Minimum number of samples required per cell type. Default: \code{5}.}

\item{min.cells}{Integer. Minimum number of cells required per cell type. Default: \code{10}.}

\item{k.sc}{Integer. Number of nearest neighbors for single-cell KNN graph. Default: \code{20}.}

\item{k.psblk}{Integer. Number of nearest neighbors for pseudobulk KNN graph. Default: \code{5}.}

\item{black.list}{(Optional) Character vector of genes to exclude from analysis.
Defaults to \code{scTypeEval@black.list}.}

\item{ncores}{Integer. Number of CPU cores to use for parallel processing. Default: \code{1}.}

\item{bparam}{Parallel backend parameter object for BiocParallel. If provided, overrides \code{ncores}.}

\item{progressbar}{Logical. Whether to display a progress bar during computation. Default: \code{FALSE}.}

\item{verbose}{Logical. Whether to print messages during execution. Default: \code{TRUE}.}
}
\value{
A data frame containing consistency evaluation results for each consistency parameter.
}
\description{
This function serves as a wrapper to compute internal validation metrics (\link[Run.Consistency]{Run.Consistency})
and Mutual Best Hit consistency (\link[Run.BestHit]{Run.BestHit}) evaluations for cell type annotations.
It supports multiple data types in the same run, including single-cell and pseudobulk data.
}
\details{
This function applies sequentially internal validation metrics (\link[Run.Consistency]{Run.Consistency})
and Mutual Best Hit consistency (\link[Run.BestHit]{Run.BestHit}) on scTypeEval object to assess the consistency of cell type annotations.
Finally, it returns a data frame with the consistency results for each consistency paramter.
}
\examples{
\dontrun{
consistency_df <- Run.scTypeEval(scTypeEval = sceval,
                                ident = "cell_type",
                                sample = "sample_id",
                                IntVal.metric = c("silhouette", "NeighborhoodPurity"), # specificy one or multiple internal validation metrics
                                BH.method = c("Mutual.Score","Mutual.Match"), # specificy Run.BestHit methods
                                data.type = c("sc", "pseudobulk"), # multiples data types will be run sequentially
                                ncores = 4)
}

}
\seealso{
\link[Run.Consistency]{Run.Consistency}, \link[Run.BestHit]{Run.BestHit}, \link[get.ConsistencyData]{get.ConsistencyData}
}
