% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTypeEval-package.R
\docType{package}
\name{scTypeEval-package}
\alias{scTypeEval}
\alias{scTypeEval-package}
\title{scTypeEval: Evaluation of cell type classifications in single-cell transcriptomics}
\description{
scTypeEval offers assessment and validation of cell type classifications in single-cell transcriptomics. In the absence of ground truth annotations, scTypeEval leverages internal validation metrics to evaluate the consistency and quality of cell type assignments, particularly across multiple biological samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/carmonalab/scTypeEval}
  \item Report bugs at \url{https://github.com/carmonalab/scTypeEval/issues}
}

}
\author{
\strong{Maintainer}: Josep Garnica \email{josep.garnicacaparros@unige.ch} (\href{https://orcid.org/0000-0001-9493-1321}{ORCID})

Authors:
\itemize{
  \item Massimo Andreatta \email{massimo.andreatta@unige.ch} (\href{https://orcid.org/0000-0002-8036-2647}{ORCID})
  \item Santiago Carmona \email{santiago.carmona@unige.ch} (\href{https://orcid.org/0000-0002-2495-0671}{ORCID})
}

}
\keyword{internal}
