% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{set.activeIdent}
\alias{set.activeIdent}
\title{Set the active ident or annotation labels for an scTypeEval object.}
\usage{
set.activeIdent(scTypeEval, ident = NULL)
}
\arguments{
\item{scTypeEval}{An \code{scTypeEval} object.}

\item{ident}{A character string specifying the column in \code{scTypeEval@metadata} to set as the active identity.}
}
\value{
The modified \code{scTypeEval} object with the active identity set.
}
\description{
This function assigns a specific cell type annotation from the metadata as the active identity
in an \code{scTypeEval} object, allowing for downstream analysis based on the selected classification.
}
\details{
The function ensures that the provided identity exists within the metadata before setting it.
If no valid identity is provided, an error is raised.
}
\examples{
\donttest{
sceval <- create.scTypeEval(count_matrix, metadata = metadata)
sceval <- set.activeIdent(sceval, ident = "cell_type")
}

}
