% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.PCA}
\alias{plot.PCA}
\title{Plot PCA Results from scTypeEval Object}
\usage{
\method{plot}{PCA}(
  scTypeEval,
  reduction.slot = "all",
  label = TRUE,
  dims = c(1, 2),
  show.legend = FALSE
)
}
\arguments{
\item{scTypeEval}{An \code{scTypeEval} object containing PCA results in the \code{reductions} slot.}

\item{reduction.slot}{Character. Name(s) of the reduction(s) to plot. If \code{"all"}
(default), all available PCA reductions in the object are plotted.}

\item{label}{Logical. Whether to add cluster labels to the PCA plot (default: \code{TRUE}).}

\item{dims}{Integer vector of length 2. The principal component (PC) dimensions to plot
(default: \code{c(1,2)}).}

\item{show.legend}{Logical. Whether to display a legend (default: \code{FALSE}).}
}
\value{
A named list of PCA plots (\link[ggplot2]{ggplot} objects) corresponding to
the PCA analyses stored in the \code{reductions} slot of an \code{scTypeEval} object
(generated by \link{add.DimReduction}). If only one reduction is selected, a single ggplot object
is returned.
}
\description{
This function visualizes Principal Component Analysis (PCA) results stored in the
\code{reductions} slot of an \code{scTypeEval} object.
}
\examples{
\dontrun{
# plot all PCA dimensional reductions
pca_plots <- plot.PCA(sceval) # scTypeEval object with previously run `add.PCA()`
}

}
\seealso{
\link{add.DimReduction}
}
